Tried to convert to 64 bit. Failed, not sure what the problem is. Probably 32 bit libs left. FreeType doesn't have a 64 bit build, look into it. Settings all setup now. Any CPU Shouldn't exist anymore.
Former-commit-id: 2e226b3579761d2d4620a202927ca8fb86d65e9f

+++ Timer/TimeManager/TimeManager.cpp
@@ -50,7 +50,7 @@ namespace CommonUtilities
 		Timer newTimer;
 		newTimer.Start();
 		myTimers.push_back(newTimer);
-		return myTimers.size() - 1;
+		return static_cast<int>(myTimers.size()) - 1;
 	}
 
 	void TimeManager::Update()
